move repo interfaces to own folder.
