+++ Src/MoneyFox.Shared.Tests/Converter/PaymentAmountConverterTests.cs
@@ -3,6 +3,7 @@ using System.Linq.Expressions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MoneyFox.Shared.Converter;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using Moq;
 using MvvmCross.Platform;
