+++ Src/MoneyFox.Shared.Tests/ViewModels/CategoryListViewModelTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.ObjectModel;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using MoneyFox.Shared.ViewModels;
 using Moq;
