+++ Src/MoneyFox.Shared.Tests/ViewModels/ModifyCategoryDialogViewModelTests.cs
@@ -4,6 +4,7 @@ using System.Linq.Expressions;
 using Cheesebaron.MvxPlugins.Settings.Interfaces;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using MoneyFox.Shared.ViewModels;
 using Moq;
