--- Src/MoneyFox.Shared/Interfaces/IData.cs
@@ -1,13 +1,0 @@
-using System.Collections.ObjectModel;
-
-namespace MoneyFox.Shared.Interfaces
-{
-    public interface IData<T>
-    {
-        /// <summary>
-        ///     All payment loaded from the database
-        /// </summary>
-        ObservableCollection<T> Data { get; set; }
-
-    }
-}
