+++ Src/MoneyFox.Shared/Interfaces/Repositories/IAccountRepository.cs
@@ -1,6 +1,6 @@
 using MoneyFox.Shared.Model;
 
-namespace MoneyFox.Shared.Interfaces
+namespace MoneyFox.Shared.Interfaces.Repositories
 {
     public interface IAccountRepository : IRepository<Account>
     {
