+++ Src/MoneyFox.Shared/Interfaces/Repositories/IRecurringPaymentRepository.cs
@@ -1,7 +1,6 @@
 using MoneyFox.Shared.Model;
-using MoneyFox.Shared.Repositories;
 
-namespace MoneyFox.Shared.Interfaces
+namespace MoneyFox.Shared.Interfaces.Repositories
 {
     public interface IRecurringPaymentRepository : IRepository<RecurringPayment>
     {
