+++ Src/MoneyFox.Shared/Interfaces/Repositories/ISelectedProperty.cs
@@ -1,0 +1,10 @@
+﻿namespace MoneyFox.Shared.Interfaces.Repositories
+{
+    public interface ISelectedProperty<T>
+    {
+        /// <summary>
+        ///     The selected Item
+        /// </summary>
+        T Selected { get; set; }
+    }
+}
