+++ Src/MoneyFox.Shared/Manager/RepositoryManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 
 namespace MoneyFox.Shared.Manager
