+++ Src/MoneyFox.Shared/Repositories/PaymentRepository.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using MoneyFox.Shared.Exceptions;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using PropertyChanged;
