+++ Src/MoneyFox.Shared/ViewModels/AbstractCategoryListViewModel.cs
@@ -3,6 +3,7 @@ using System.Globalization;
 using System.Linq;
 using MoneyFox.Shared.Groups;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using MoneyFox.Shared.Resources;
 using MvvmCross.Core.ViewModels;
