+++ Src/MoneyFox.Shared/ViewModels/BalanceViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using MoneyFox.Shared.Helpers;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Interfaces.ViewModels;
 using MoneyFox.Shared.Model;
 using MvvmCross.Core.ViewModels;
