+++ Src/MoneyFox.Shared/ViewModels/CategoryListViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using PropertyChanged;
