+++ Src/MoneyFox.Shared/ViewModels/MainViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using MvvmCross.Binding.ExtensionMethods;
 using MvvmCross.Core.ViewModels;
