+++ Src/MoneyFox.Shared/ViewModels/ModifyCategoryDialogViewModel.cs
@@ -2,6 +2,7 @@ using System;
 using System.Linq;
 using MoneyFox.Shared.Helpers;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using MoneyFox.Shared.Resources;
 using MvvmCross.Core.ViewModels;
