+++ Src/MoneyFox.Shared/ViewModels/PaymentListBalanceViewModel.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using MoneyFox.Shared.Helpers;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 
 namespace MoneyFox.Shared.ViewModels
