+++ Src/MoneyFox.Shared/ViewModels/SelectCategoryListViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Messages;
 using MoneyFox.Shared.Model;
 using MvvmCross.Core.ViewModels;
