+++ Src/MoneyFox.Shared/ViewModels/StatisticCashFlowViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using MoneyFox.Shared.Helpers;
 using MoneyFox.Shared.Interfaces;
+using MoneyFox.Shared.Interfaces.Repositories;
 using MoneyFox.Shared.Model;
 using MoneyFox.Shared.StatisticDataProvider;
 using OxyPlot;
